Todos los arreglos son double, la biblioteca Matlab se modifico

+++ Datos_Propiedades.cs
@@ -11,7 +11,7 @@ namespace Lanzamientos_Espaciales_Exitosos
         #region Variables
         public string rutaArchivo { get; set; }
         public string nombreDatos { get; set; }
-        public int [] alturas { get; set; }
+        public double [] alturas { get; set; }
         public double promedio { get; set; }
         public double maxValFunDensi { get; set; }
         public double minValFunDensi { get; set; }
